Added banana bread recipe page + formatting changes
Created the banana bread recipe page with the required content. I have updated the index page so it has a link to the banana bread page. I have made it so that all text is on a single line (in the file) rather than manually splitting the text over multiple lines